fixing build erorrs
