+++ src/app/components/ShopSection.tsx
@@ -1,4 +1,4 @@
-import { ShopItem } from '../pages/Dashboard';
+import { ShopItem } from '@/app/dashboard/Dashboard';
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from './ui/card';
 import { Button } from './ui/button';
 import { Badge } from './ui/badge';
